Initial gcc4 warning cleanup (and some minor policing)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@8063 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/common/memcpy.c
@@ -95,7 +95,7 @@ _DEFUN (memcpy, (dst0, src0, len0),
           *aligned_dst++ = *aligned_src++;
           *aligned_dst++ = *aligned_src++;
           *aligned_dst++ = *aligned_src++;
-          len -= BIGBLOCKSIZE;
+          len -= (unsigned int)BIGBLOCKSIZE;
         }
 
       /* Copy one long word at a time if possible.  */
